Fix many of the intcap_t bitwise operation warnings by either:
1) using the new q{Get,Set,Clear}LowPointerBits functions in QtBase
2) using __builtin_align_{up, down}
3) replacing the intptr_t with an integer type if that is the intention

+++ Source/JavaScriptCore/bytecode/PutByIdFlags.h
@@ -30,7 +30,7 @@
 
 namespace JSC {
 
-enum PutByIdFlags : intptr_t {
+enum PutByIdFlags : int {
     PutByIdNone = 0,
 
     // This flag indicates that the put_by_id is direct. That means that we store the property without
